requirements.txt
astunparse==1.6.3
attrs==22.2.0
build==0.10.0
click==8.1.3
chromadb==0.3.26
datasets==2.10.1
fpdf2==2.7.4
gradio==3.35.2
gnureadline==8.1.2
langchain==0.0.265
llama-index==0.8.2.post1
lorem==0.1.1
matplotlib==3.7.1
nltk==3.8.1
numpy==1.24.2
openai==0.27.2
pandasql==0.7.3
pdf2image==1.16.3
pylint==2.17.2
pytest==7.4.0
pytest_cov==4.1.0
PyMuPDF==1.22.5
pypdf==3.9.0
PyYAML==6.0
rank_bm25==0.2.2
scikit-learn==1.2.2
sentence-transformers==2.2.2
tiktoken==0.3.3
torch==2.0.0
tqdm==4.65.0
transformers==4.30.0
twine==4.0.2
